feature RegisterFormValidator
